Improvements to the frontend

+++ frontend/src/utils/to-do/delete.ts
@@ -3,12 +3,18 @@ import IOperations from 'interfaces/to-do/Operations';
 import { doFetch } from "utils/fetch";
 
 const _delete = async ({ sessionId, title, refresh }: IOperations) => {
-    if (!title) return;
+    const modal = document.getElementById('delete-modal');
+    const deleteButton = document.getElementById('delete-modal-accept');
 
-    // TODO: Display "Are you sure" message
+    if (!title || !modal || !deleteButton) return;
 
-    await doFetch({ url: 'to-dos/', method: 'delete', body: { sessionId, title } });
-    await refresh();
+    modal.style.display = 'block';
+
+    deleteButton.onclick = async () => {
+        await doFetch({ url: 'to-dos/', method: 'delete', body: { sessionId, title } });
+        modal.style.display = 'none';
+        await refresh();
+    }
 }
 
 export default _delete;
